requirements.txt
aes==1.0.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-MQTT==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mqtt==0.0.1
paho-mqtt==1.6.1
pycryptodomex==3.14.0
python-dotenv==0.19.2
Werkzeug==2.0.2
